fix: moved gauzy range picker to gauzy filter
