+++ apps/gauzy/src/app/@shared/timesheet/gauzy-filters/gauzy-filters.component.scss
