+++ apps/gauzy/src/app/@theme/components/header/selectors/date-range-picker/arrow/strategies/index.ts
