+++ apps/gauzy/src/app/pages/reports/apps-urls-report/apps-urls-report/apps-urls-report.component.html
@@ -17,11 +17,11 @@
 				</ng-container>
 			</div>
 			<div class="filters">
-				<ngx-gauzy-range-picker
+				<ngx-gauzy-filters
 					[filters]="logRequest"
 					[saveFilters]="false"
 					(filtersChange)="filtersChange($event)"
-				></ngx-gauzy-range-picker>
+				></ngx-gauzy-filters>
 			</div>
 		</div>
 	</nb-card-header>
