+++ apps/gauzy/src/app/pages/reports/client-budgets-report/client-budgets-report.module.ts
@@ -15,7 +15,7 @@ import { SharedModule } from '../../../@shared/shared.module';
 import { ProgressStatusModule } from '../../../@shared/progress-status/progress-status.module';
 import { HeaderTitleModule } from '../../../@shared/components/header-title/header-title.module';
 import { DateRangeTitleModule } from '../../../@shared/components/date-range-title/date-range-title.module';
-import { GauzyRangePickerModule } from '../../../@shared/timesheet/gauzy-range-picker/gauzy-range-picker.module';
+import { GauzyFiltersModule } from '../../../@shared/timesheet/gauzy-filters/gauzy-filters.module';
 
 @NgModule({
 	declarations: [ClientBudgetsReportComponent],
@@ -33,7 +33,7 @@ import { GauzyRangePickerModule } from '../../../@shared/timesheet/gauzy-range-p
 		ProgressStatusModule,
 		HeaderTitleModule,
 		DateRangeTitleModule,
-   		GauzyRangePickerModule
+   		GauzyFiltersModule
 	]
 })
 export class ClientBudgetsReportModule {}
