+++ apps/gauzy/src/app/pages/reports/client-budgets-report/client-budgets-report/client-budgets-report.component.html
@@ -12,14 +12,14 @@
 	</nb-card-header>
 	<nb-card-body>
 		<div class="filters mt-4 mb-4">
-			<ngx-gauzy-range-picker
+			<ngx-gauzy-filters
 				[hasActivityLevelFilter]="false"
 				[hasSourceFilter]="false"
 				[hasLogTypeFilter]="false"
 				[saveFilters]="false"
 				[filters]="logRequest"
 				(filtersChange)="filtersChange($event)"
-			></ngx-gauzy-range-picker>
+			></ngx-gauzy-filters>
 		</div>
 		<div
 			class="weekly-logs row-table"
