+++ apps/gauzy/src/app/pages/reports/expenses-report/expenses-report.module.ts
@@ -15,7 +15,7 @@ import { LineChartModule } from '../../../@shared/report/charts/line-chart/line-
 import { HeaderTitleModule } from '../../../@shared/components/header-title/header-title.module';
 import { SharedModule } from '../../../@shared/shared.module';
 import { DateRangeTitleModule } from '../../../@shared/components/date-range-title/date-range-title.module';
-import { GauzyRangePickerModule } from '../../../@shared/timesheet/gauzy-range-picker/gauzy-range-picker.module';
+import { GauzyFiltersModule } from '../../../@shared/timesheet/gauzy-filters/gauzy-filters.module';
 
 @NgModule({
 	declarations: [ExpensesReportComponent],
@@ -33,7 +33,7 @@ import { GauzyRangePickerModule } from '../../../@shared/timesheet/gauzy-range-p
 		LineChartModule,
 		HeaderTitleModule,
 		DateRangeTitleModule,
-    	GauzyRangePickerModule
+    	GauzyFiltersModule
 	]
 })
 export class ExpensesReportModule {}
