+++ apps/gauzy/src/app/pages/reports/manual-time/manual-time.module.ts
@@ -15,7 +15,7 @@ import {
 } from '@nebular/theme';
 import { HeaderTitleModule } from '../../../@shared/components/header-title/header-title.module';
 import { DateRangeTitleModule } from '../../../@shared/components/date-range-title/date-range-title.module';
-import { GauzyRangePickerModule } from '../../../@shared/timesheet/gauzy-range-picker/gauzy-range-picker.module';
+import { GauzyFiltersModule } from '../../../@shared/timesheet/gauzy-filters/gauzy-filters.module';
 import { OrganizationModule } from "../../../@shared/report/organization/organization.module";
 
 @NgModule({
@@ -33,7 +33,7 @@ import { OrganizationModule } from "../../../@shared/report/organization/organiz
 		NbBadgeModule,
 		HeaderTitleModule,
 		DateRangeTitleModule,
-		GauzyRangePickerModule,
+		GauzyFiltersModule,
 		OrganizationModule,
 	],
 })
