+++ apps/gauzy/src/app/pages/reports/manual-time/manual-time/manual-time.component.html
@@ -16,14 +16,14 @@
     </nb-card-header>
     <nb-card-body>
         <div class="filters">
-            <ngx-gauzy-range-picker
+            <ngx-gauzy-filters
                 [filters]="logRequest"
                 [saveFilters]="false"
                 [hasLogTypeFilter]="false"
                 [hasSourceFilter]="false"
                 [hasActivityLevelFilter]="false"
                 (filtersChange)="filtersChange($event)"
-            ></ngx-gauzy-range-picker>
+            ></ngx-gauzy-filters>
             <nb-select placeholder="Action" class="action-select">
                 <nb-option value="0">Action</nb-option>
             </nb-select>
