+++ apps/gauzy/src/app/pages/reports/payment-report/payment-report/payment-report.component.html
@@ -21,14 +21,14 @@
 	</nb-card-header>
 	<nb-card-body>
 		<div class="filters mt-4 mb-4">
-			<ngx-gauzy-range-picker
+			<ngx-gauzy-filters
 				[filters]="logRequest"
 				[hasActivityLevelFilter]="false"
 				[hasSourceFilter]="false"
 				[hasLogTypeFilter]="false"
 				[saveFilters]="false"
 				(filtersChange)="filtersChange($event)"
-			></ngx-gauzy-range-picker>
+			></ngx-gauzy-filters>
 		</div>
 		<div class="daily-time-report">
 			<nb-card>
