+++ apps/gauzy/src/app/pages/reports/project-budgets-report/project-budgets-report/project-budgets-report.component.html
@@ -16,14 +16,14 @@
 	</nb-card-header>
 	<nb-card-body>
 		<div class="filters mt-4 mb-4">
-			<ngx-gauzy-range-picker
+			<ngx-gauzy-filters
 				[filters]="logRequest"
 				[hasActivityLevelFilter]="false"
 				[hasSourceFilter]="false"
 				[hasLogTypeFilter]="false"
 				[saveFilters]="false"
 				(filtersChange)="filtersChange($event)"
-			></ngx-gauzy-range-picker>
+			></ngx-gauzy-filters>
 		</div>
 		<div
 			class="weekly-logs row-table"
