+++ apps/gauzy/src/app/pages/reports/time-limit-report/time-limit-report.module.ts
@@ -16,7 +16,7 @@ import { LineChartModule } from '../../../@shared/report/charts/line-chart/line-
 import { SharedModule } from '../../../@shared/shared.module';
 import { HeaderTitleModule } from '../../../@shared/components/header-title/header-title.module';
 import { DateRangeTitleModule } from '../../../@shared/components/date-range-title/date-range-title.module';
-import { GauzyRangePickerModule } from '../../../@shared/timesheet/gauzy-range-picker/gauzy-range-picker.module';
+import { GauzyFiltersModule } from '../../../@shared/timesheet/gauzy-filters/gauzy-filters.module';
 
 @NgModule({
 	declarations: [TimeLimitReportComponent],
@@ -34,7 +34,7 @@ import { GauzyRangePickerModule } from '../../../@shared/timesheet/gauzy-range-p
 		ProgressStatusModule,
 		HeaderTitleModule,
 		DateRangeTitleModule,
-    	GauzyRangePickerModule
+    	GauzyFiltersModule
 	]
 })
 export class TimeLimitReportModule {}
