+++ apps/gauzy/src/app/pages/reports/time-limit-report/time-limit-report/time-limit-report.component.html
@@ -16,11 +16,11 @@
 	</nb-card-header>
 	<nb-card-body>
 		<div class="filters mt-4 mb-4">
-			<ngx-gauzy-range-picker
+			<ngx-gauzy-filters
         		[filters]="logRequest"
         		[saveFilters]="false"
         		(filtersChange)="filtersChange($event)">
-			</ngx-gauzy-range-picker>
+			</ngx-gauzy-filters>
 		</div>
 		<div class="table">
 			<div
