+++ apps/gauzy/src/app/pages/reports/weekly-time-reports/weekly-time-reports.module.ts
@@ -9,7 +9,7 @@ import { LineChartModule } from '../../../@shared/report/charts/line-chart/line-
 import { DailyStatisticsModule } from '../../../@shared/report/daily-statistics/daily-statistics.module';
 import { HeaderTitleModule } from '../../../@shared/components/header-title/header-title.module';
 import { DateRangeTitleModule } from '../../../@shared/components/date-range-title/date-range-title.module';
-import { GauzyRangePickerModule } from '../../../@shared/timesheet/gauzy-range-picker/gauzy-range-picker.module';
+import { GauzyFiltersModule } from '../../../@shared/timesheet/gauzy-filters/gauzy-filters.module';
 
 @NgModule({
 	declarations: [WeeklyTimeReportsComponent],
@@ -25,7 +25,7 @@ import { GauzyRangePickerModule } from '../../../@shared/timesheet/gauzy-range-p
 		NbCardModule,
 		HeaderTitleModule,
 		DateRangeTitleModule,
-		GauzyRangePickerModule,
+		GauzyFiltersModule,
 	],
 })
 export class WeeklyTimeReportsModule {}
