+++ apps/gauzy/src/app/pages/reports/weekly-time-reports/weekly-time-reports/weekly-time-reports.component.html
@@ -22,11 +22,11 @@
 				</ng-container>
 			</div>
 			<div class="filters">
-				<ngx-gauzy-range-picker
+				<ngx-gauzy-filters
 					[filters]="logRequest"
 					[saveFilters]="false"
 					(filtersChange)="filtersChange($event)"
-				></ngx-gauzy-range-picker>
+				></ngx-gauzy-filters>
 			</div>
 		</div>
 	</nb-card-header>
